chore: generate Expo's tRPC types using tsc instead of xtrpc (#231)
I've been having more and more problems with xtrpc not generating types
correctly / erroring when the input and output types become nontrivial.

Saw this approach recommended in some other posts and it seems to work a
lot better.

Blocker for Helper agent (it changes the search endpoint in a way that
xtrpc refuses to parse).